refactor(src): Move view sub-packages out of view.gui, into view
