+++ src/main/java/edu/ntnu/idi/idatt/App.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idi.idatt;
 
-import edu.ntnu.idi.idatt.view.gui.container.AppView;
+import edu.ntnu.idi.idatt.view.container.AppView;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.scene.layout.BorderPane;
